requirements.txt
numpy==2.1.3
pandas==2.2.3
streamlit==1.40.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
PyYAML==6.0
openpyxl==3.1.5
